+++ src/interactive/mod.rs
@@ -61,7 +61,7 @@ pub fn show(
     client: Client,
     connection: Connection,
     broker: &Broker,
-    subscribe_topic: String,
+    subscribe_topic: Vec<String>,
 ) -> anyhow::Result<()> {
     let mqtt_thread = mqtt_thread::MqttThread::new(client, connection, subscribe_topic)?;
     let mut app = App::new(broker, mqtt_thread);